refactor: extract expand/refine hooks from UniverseBuilderPage (#3391)
Expand + refine already lived in useUniverseExpand, so the page-body reduction
the issue asks for landed on the flows still inline:

- useUniverseTabs: URL-backed tab/sub-bucket state and the two effects that
  strip a ?tab=/?bucket= the current categories no longer support
- useUniverseBucketActions: generate-in-bucket, promote-to-canon, auto-sort,
  each with its own ref+state in-flight gate and stale-snapshot merge rules
- the render-completion handler copy-pasted between the trunk tabs and the
  Other tab is now one shared callback

Pure refactor - same API calls, same effect ordering, same markup. The page is
530 lines (was 778); useUniverseTabs picks up 11 tests.
