+++ .changelog/NEXT.md
@@ -25,6 +25,7 @@
 - **A canon trunk's wiring travels by context, not a 36-prop bag.** `KindSection` inside `UniverseCanonSection.jsx` took every handler, spinner id, and job map as an individually threaded prop, so adding one capability meant editing the call site, the signature, and its defaults. The bag is now published once per kind through `KindSectionContext` and read where it's used, and the inline add-entry mini-form moved to `CanonAddEntryForm.jsx` as a controlled leaf — the typed draft stays with the section, because the toolbar "Add" button is a toggle and collapsing the form has to bring the draft back on reopen. `KindSection` takes zero props; behavior, markup, and classes are unchanged. (#3388)
 - **The Round editor is a composition, not a 1,326-line component.** `RoundEditor.jsx` drops to 164 lines of page shell — header chrome plus the pick between its three body surfaces — with the state split into four named hooks (`useRoundViewParams` for the `?mode`/`?stack`/`?analyze` views, `useRoundDraft` for load/save/refresh-from-template/AI-merge and the `baseDirty` saved-state gate, `useRoundRows` for section/layer/reference/proposed-part edits, `useRoundPartners` for the quodlibet pick list) and the markup into five components in `client/src/components/songs/` (`RoundEditForm`, `RoundReadView`, `RoundSheetMusic`, `RoundReferenceCard`, `RoundBuiltInBanner`). The draft rules the page, hooks, and form all shared — server caps, the temp-id contract, the tempo band, and the save patch builder — are now `client/src/lib/roundDraft.js`, and the host-anchored TikTok embed-id helpers moved to `utils/urlNormalize.js` with tests. Behavior is unchanged. (#3389)
 - `agentLifecycle.js` imported `getAgent` aliased as `getAgentRecord`, shadowing the real transcript-free reader that was split out in #3368 for exactly this purpose. Both call sites read only `.status`/`.metadata`, but the alias made them hydrate the record: reading and line-splitting a completed run's whole `output.txt` (megabytes on a long TUI run) and potentially rewriting `metadata.json` via `repairCodexTaskSummary`. Now imports the real `getAgentRecord`, read once and shared.
+- **Universe Builder is a composition shell again.** `UniverseBuilderPage.jsx` drops from 778 to ~530 lines: the URL-backed tab/sub-bucket state (plus the two effects that strip a `?tab=`/`?bucket=` the current categories no longer support) is now `useUniverseTabs`, and generate-in-bucket / promote-to-canon / auto-sort — each with its own in-flight gate and stale-snapshot merge rules — is now `useUniverseBucketActions`. The render-completion handler that was copy-pasted between the trunk tabs and the Other tab is one shared callback. Expand/refine already lived in `useUniverseExpand`, so it stayed put. Behavior, markup, and API calls are unchanged; `useUniverseTabs` picks up 11 tests. (#3391)
 
 ## Better audit 2026-08-03 (UX, City scene, game/music/writing focus) — PRs #3401–#3404
